Make every unit definition independent
The goal of this commit is that each unit should be defined purely by
its dimension, and an essentially "magic" number, with no explicit
reference to any other unit.  They all enter on equal footing.  We rely
on the _unit tests_ --- which we do not change --- to make sure that
every unit has the correct relationship with every other unit.

This simplifies the logical structure of the library, and reduces
entanglements.  It may also slightly improve compile time performance.

The downside is that any code that was relying on _transitive includes_
to bring in some of the units that it needed will break.  The fix is
very simple: include the file directly for every unit that you use.

Finally, we completely abandon user-friendly definitions and aliases in
our unit definitions.  Since these units are included directly with the
library, we have enough test coverage to be perfectly confident that we
got them right. Therefore, we spell out the concrete `Dimension<...>`
and `Magnitude<...>` types directly.  End users must never do this in
their code, so we include a stark extra warning with every instance.
This direct approach should be the easiest possible for compilers to
handle, which will likely help #560, and may even fix it.

Fixes #451.