Rename internal alias for corresponding quantity
This is just internal to the library, but it was still a "names differ
by only `T` suffix" situtaion, so we wanted to fix it.

In the unlikely event that any external users are using this, they can
simply find-and-replace and update to the new name.  The number of uses
should be very small at most.

Helps #86.